Issue release message after response. (#15)
Send release message after receiving query response.

Must do this in order to reacquire agency and read an up-to-date version of the ledger.
This fixes an issue where the ledger state tracked by the Query OTP process wouldn't be 
updated after establishing a connection. The result was that queries would always return the same value.

This PR refactors Query states and events. The goal is to only acquire agency prior to 
submitting a query to the node. This ensures that the query will be performed against 
the most up-to-date version of the ledger.